% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AG_of_TC.R
\name{AG_of_TC}
\alias{AG_of_TC}
\title{Aantonen of er meer purines of pyrimidines zijn}
\usage{
AG_of_TC(sequentie)
}
\arguments{
\item{sequentie, }{Vul hier de DNA sequentie in}
}
\value{
Een tekst met het antwoord van de functie
}
\description{
Aantonen of er meer purines of pyrimidines zijn
}
\examples{
sequentie<- "ACGTTGCA"
AG_of_TC(sequentie)
}
