% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengte_sequentie.R
\name{lengte_sequentie}
\alias{lengte_sequentie}
\title{Lengte van de sequentie teruggeven}
\usage{
lengte_sequentie(sequentie)
}
\arguments{
\item{sequentie, }{Vul hier de DNA sequentie in}
}
\value{
Een nummer wat de hoeveelheid nucleotide in de sequentie uitdrukt
}
\description{
Lengte van de sequentie teruggeven
}
\examples{
sequentie<- "ACGTTGCA"
lengte_sequentie(sequentie)
}
