% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_GC.R
\name{percentage_GC}
\alias{percentage_GC}
\title{Berekenen van de hoeveelheid GC in een DNA sequentie}
\usage{
percentage_GC(sequentie)
}
\arguments{
\item{sequentie, }{Vul hier de DNA sequentie in}
}
\value{
De hoeveelheid GC in percentage
}
\description{
Berekenen van de hoeveelheid GC in een DNA sequentie
}
\examples{
sequentie<- "ACGTTGCA"
percentage_GC(sequentie)
}
