% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellen_nucleotides.R
\name{tellen_nucleotides}
\alias{tellen_nucleotides}
\title{Tellen van de hoeveelheid van elke nucleotide}
\usage{
tellen_nucleotides(sequentie)
}
\arguments{
\item{sequentie, }{Vul hier de DNA sequentie in}
}
\value{
Een overzicht van de hoeveelheid verschillende nucleotide
}
\description{
Tellen van de hoeveelheid van elke nucleotide
}
\examples{
sequentie<- "ACGTTGCA"
tellen_nucleotides(sequentie)
}
